Generate static library
Amdllpc requires to link spvgen statically, so add build info to
generate static library. The static will depend on amdllpc build.
Add an object library to compile source files of spvgen, static and
dynamic depend it.

Add two internal functions, they are not exported and they are called
when process is initialized and terminated with using static library.
Add a interface "FinalizeSpvgen" which clean up resouces.